Cleanup old libs.tickets
