--- bot/libs/tickets/__init__.py
@@ -1,1 +1,0 @@
-"""Package for various utils, views, and others dedicated for handling tickets"""
